+++ src/main/kotlin/com/github/andronikusgametech/ijreplaymod/actions/CodingReplayErrorDialogue.kt
@@ -1,15 +1,14 @@
-package com.github.andronikusgametech.ijreplaymod.util
+package com.github.andronikusgametech.ijreplaymod.actions
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogWrapper
 import java.awt.BorderLayout
-import java.awt.Dimension
 import javax.swing.JComponent
 import javax.swing.JLabel
 import javax.swing.JPanel
 
 class CodingReplayErrorDialogue(
-    private val project: Project,
+    project: Project,
     private val error: String
 ): DialogWrapper(project) {