+++ src/main/kotlin/com/github/andronikusgametech/ijreplaymod/model/Keyframe.kt
@@ -6,5 +6,9 @@ data class Keyframe(
     @Tag
     public var order: Int = 0,
     @Tag
+    public var id: Int = 0,
+    @Tag
+    public var label: String = "",
+    @Tag
     public var text: String = ""
 )